jwt access token user login protection routes user verify jwt token middleware, protect routes contact, get all contacts, create new contact, update and delete contact

+++ controllers/contactcontroller.js
@@ -2,35 +2,36 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id });
     res.status(200).json({ contacts });
 });
 
 
 //@desc Create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     console.log("The request body is: ", req.body);
     const { name, email, phone } = req.body;
     if (!name || !email || !phone) {
         res.status(400);
         throw new Error("All fields are mandatory !");
-    }
+    };
 
     const contact = await Contact.create({
         name,
         email,
         phone,
+        user_id: req.user.id
     });
     res.status(201).json( contact );
 });
 
 //@desc Get contact
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
@@ -42,12 +43,17 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Update contacts
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
      if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
     }
 
     const updateContact = await Contact.findByIdAndUpdate(
@@ -61,12 +67,17 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete contacts
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
      if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
     }
     // await Contact.remove();
     await contact.deleteOne(); // or use Contact.findByIdAndDelete(req.params.id)
